Re-enabled ignored tests
Tests:
  * SQL - Backend issues has been resolved, this now passes
  * SSH - Edge issues has been resolved, this fails due to an issue in
runtime where the registry variable doesn't set the --host flag
